Cloud Function for making separate RGB fits files
This CF accepts a path to a CR2 file in the storage bucket and will
create 3 separate FITS files, one for each red, green, and blue channel,
and then upload the results back to a bucket. The FITS headers from the
FITS file corresponding the CR2 file are also added to each colored
FITS.

The CF also allows for processing options to be passed to the underlying
[rawpy](https://github.com/letmaik/rawpy) module via a json POST.

Note: This currently uploads the files to a temporary bucket and is not
necessarily optimized. WIP.