Make Observation PSC - DataFlow
This creates a template for a dataflow job. The job accepts a location
of an observation sequence as stored in the `panoptes-detected-sources`
bucket. For now this is a full bucket path but soon will be just the
`sequence_id`.  Currently adds all the CSV files together, then filters
the sources that don't appear in all the frames. Also computes the
normalized data so the similar stars finder can be added.

Todo:
* Accept just a `sequence_id`.
* Figure out best way to call.
* Add in the `frameThreshold` param.
* Do Similar Stars step